Rice_Grain_Size_Calculator/requirements.txt
opencv-python-headless
numpy
pandas
matplotlib
streamlit
Pillow
ultralytics
streamlit-image-comparison